requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chromedriver-autoinstaller==0.6.4
click==8.2.0
click-default-group==1.2.4
colorama==0.4.6
comm==0.2.2
condense-json==0.1.3
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
distro==1.9.0
executing==2.1.0
fastapi==0.115.12
filelock==3.18.0
fire==0.7.0
fonttools==4.55.3
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.32.3
idna==3.10
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.1
json5==0.12.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
llm==0.25
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistletoe==1.4.0
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.0
openai==1.82.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==5.29.5
psutil==6.1.0
pure_eval==0.2.3
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-frontmatter==1.1.0
python-multipart==0.0.20
python-ulid==3.0.0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
selenium==4.33.0
sentence-transformers==4.1.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
split-markdown4gpt==1.0.9
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
syntok==1.4.4
tabulate==0.9.0
termcolor==3.1.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2024.2
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.4.0
uv==0.7.5
uvicorn==0.34.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
wsproto==1.2.0
